rpk/cli: Add 'rpk topic describe', depr. 'rpk api topic describe'
